+++ src/ui.rs
@@ -8,6 +8,7 @@ mod builder;
 mod dbus;
 mod dialog_about;
 mod dialog_archive_prefix;
+mod dialog_delete_archive;
 mod dialog_device_missing;
 mod dialog_encryption_password;
 mod dialog_exclude;